requirements.txt
django==1.9.3
psycopg2==2.6.1
django-cors-headers==1.1.0
djangorestframework==3.3.2
django-filter==0.13.0
djangorestframework-filters==0.7.0
django-admin-bootstrapped==2.5.7
simplejson==3.8.2
pillow==3.1.1